Add redirect hook and refactor ProtectedRoute
